3.1.2: Auto stash before merge of "LWJGL-3.1.2" and "origin/LWJGL-3.1.2"

+++ src/omnikryptec/util/Instance.java
@@ -109,7 +109,7 @@ public class Instance {
 		return getDisplayManager().getDeltaTime();
 	}
 	
-	public static final long getFPS(){
+	public static final double getFPS(){
 		return getDisplayManager().getFPS();
 	}
